fixes in Oracle data types; added Check decorator import in index.ts; fixed falling test;

+++ test/github-issues/1427/entity/Post.ts
@@ -6,7 +6,7 @@ export class Post {
     @PrimaryGeneratedColumn()
     id: number;
 
-    @Column({type: "double"})
+    @Column({type: "float"})
     qty: number;
 
 }
